+++ apps/recipe/src/services/recipe.ts
@@ -1,11 +1,9 @@
-import axios from "axios";
 import type { Recipe } from "@/types/recipe";
-
-const API_BASE_URL = process.env.CMS_URL;
+import cms from "./cms";
 
 export const getRecipes = async (): Promise<Recipe[]> => {
   try {
-    const response = await axios.get<Recipe[]>(`${API_BASE_URL}/recipes`);
+    const response = await cms.get<Recipe[]>("/recipes");
     return response.data;
   } catch (error) {
     console.error(error);
@@ -15,7 +13,7 @@ export const getRecipes = async (): Promise<Recipe[]> => {
 
 export const getRecipeById = async (id: string): Promise<Recipe | null> => {
   try {
-    const response = await axios.get<Recipe>(`${API_BASE_URL}/recipes/${id}`);
+    const response = await cms.get<Recipe>(`/recipes/${id}`);
 
     return response.data;
   } catch (error) {