+++ packages/components/login-form/index.ts
@@ -1,0 +1,4 @@
+import LoginPage from "./src/LoginPage";
+export * from "./src/LoginForm";
+
+export default LoginPage;